move `MonolithDiscoverer` trait and implementations to `ott-common` crate
