+++ Cargo.toml
@@ -33,6 +33,7 @@ tokio-tungstenite = "0.21.0"
 tokio-util = "0.7.8"
 tracing = "0.1.40"
 tracing-subscriber = "0.3.17"
+trust-dns-resolver = { version = "0.22.0", features = ["system-config"] }
 tungstenite = "0.21.0"
 typeshare = "1.0.0"
 url = "2.3.1"
