+++ crates/ott-common/Cargo.toml
@@ -4,15 +4,22 @@ version = "0.10.0"
 edition = "2021"
 
 [dependencies]
+anyhow.workspace = true
 async-trait.workspace = true
 futures-util.workspace = true
 hyper.workspace = true
 hyper-util.workspace = true
 http-body-util.workspace = true
+ott-balancer-protocol.workspace = true
 pin-project.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+serde_path_to_error.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 tokio.workspace = true
 tokio-tungstenite.workspace = true
 tokio-util.workspace = true
+trust-dns-resolver.workspace = true
 tungstenite.workspace = true
+url.workspace = true
