+++ crates/ott-common/src/discovery/dns.rs
