+++ crates/ott-common/src/discovery/harness.rs
