+++ crates/ott-common/src/discovery/manual.rs
