+++ crates/ott-common/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod discovery;
 pub mod websocket;
